refactor(blocks): move virutal keyboard controller to affine components (#8663)
### What Changes:
- Move `VirtualKeyboardController` to `affine-components` for common usage
