+++ packages/affine/components/src/virtual-keyboard/index.ts
@@ -1,0 +1,1 @@
+export * from './controller.js';
